move symbols to shared
