+++ src/tokens/apex/symbols.js
@@ -1,17 +1,5 @@
 const { createToken } = require('../_shared')
 
-const LBrace = createToken({
-  name: 'LBrace',
-  pattern: /\(/,
-  label: "'('",
-})
-
-const RBrace = createToken({
-  name: 'RBrace',
-  pattern: /\)/,
-  label: "')'",
-})
-
 const LCurly = createToken({
   name: 'LCurly',
   // using a string literal to get around a bug in regexp-to-ast
@@ -26,78 +14,16 @@ const RCurly = createToken({
   label: "'}'",
 })
 
-const LSquare = createToken({
-  name: 'LSquare',
-  pattern: /\[/,
-  label: "'['",
-})
-
-const RSquare = createToken({
-  name: 'RSquare',
-  pattern: /]/,
-  label: "']'",
-  pop_mode: true,
-})
-
-const Less = createToken({
-  name: 'Less',
-  pattern: /</,
-  label: "'<'",
-})
-
 const LessEquals = createToken({
   name: 'LessEquals',
   pattern: /<=/,
   label: "'<='",
 })
 
-const Greater = createToken({
-  name: 'Greater',
-  pattern: />/,
-  label: "'>'",
-})
-
 const GreaterEquals = createToken({
   name: 'GreaterEquals',
   pattern: />=/,
   label: "'>='",
-})
-
-const Dot = createToken({
-  name: 'Dot',
-  pattern: /\./,
-  label: "'.'",
-})
-
-const Comma = createToken({
-  name: 'Comma',
-  pattern: /,/,
-  label: "','",
-})
-
-const SemiColonWithFollowEmptyLine = createToken({
-  name: 'SemiColonWithFollowEmptyLine',
-  pattern: /;[ \t]*(\r\n|\r[^\n]|\n)[ \t]*(\r\n|\r|\n)/,
-  label: "';'",
-  line_breaks: true,
-})
-
-const SemiColon = createToken({
-  name: 'SemiColon',
-  pattern: /;/,
-  label: "';'",
-})
-
-const Colon = createToken({
-  name: 'Colon',
-  pattern: /:/,
-  label: "':'",
-})
-
-const Equals = createToken({
-  name: 'Equals',
-  pattern: /=/,
-  label: "'='",
 })
 
 const EqualsEquals = createToken({
@@ -227,14 +153,8 @@ const DashEquals = createToken({
 })
 
 module.exports = {
-  Dot,
-  Comma,
-  SemiColonWithFollowEmptyLine,
-  SemiColon,
-  Colon,
   EqualsEquals,
   ExclamationmarkEquals,
-  Equals,
   PlusPlus,
   PlusEquals,
   Plus,
@@ -247,16 +167,10 @@ module.exports = {
   OrOr,
   OrEquals,
   Or,
-  LBrace,
-  RBrace,
   LCurly,
   RCurly,
-  LSquare,
-  RSquare,
   LessEquals,
-  Less,
   GreaterEquals,
-  Greater,
   At,
   StarEquals,
   Star,
